# Improve performance of ParseUfloat