--- a/bytesconv_test.go
+++ b/bytesconv_test.go
@@ -239,6 +239,7 @@ func TestParseUfloatSuccess(t *testing.T) {
 	testParseUfloatSuccess(t, "123456789123456789.987654321", 123456789123456789.987654321)
 	testParseUfloatSuccess(t, "1.23456789123456789987654321", 1.23456789123456789987654321)
 	testParseUfloatSuccess(t, "340282346638528859811704183484516925440", 340282346638528859811704183484516925440)
+	testParseUfloatSuccess(t, "00000000000000000001", 00000000000000000001)
 }
 
 func TestParseUfloatError(t *testing.T) {
@@ -268,6 +269,10 @@ func TestParseUfloatError(t *testing.T) {
 
 	// missing exponent
 	testParseUfloatError(t, "123534e")
+
+	// negative number
+	testParseUfloatError(t, "-1")
+	testParseUfloatError(t, "-Inf")
 }
 
 func testParseUfloatError(t *testing.T, s string) {