@erikdubbelboer 2024-09-21
This is the first time I'm noticing that our `ParseUfloat` actually returns the wrong values for very high numbers. Even though it's faster than `strconv.ParseFloat`. I still think we should switch to `strconv.ParseFloat` as being incorrect like this is very bad. What do you think?
```
1234e23
ParseUfloat:        123399999999999981930938368.000000
strconv.ParseFloat: 123399999999999999110807552.000000

1.234e+32
ParseUfloat:        123399999999999984608249181634560.000000
strconv.ParseFloat: 123400000000000002622647691116544.000000

123456789123456789.987654321
ParseUfloat:        17502027876.294655
strconv.ParseFloat: 123456789123456784.000000
```