@ksw2000 2024-09-22
Actually, I encountered this problem as well. I believe the reason is that we use a `uint64` variable to represent a floating-point number. The largest value of `uint64` is 1.8×10¹⁹. Therefore, the third example you provided, 123456789123456789.987654321, which has 27 digits, causes the `uint64` value to overflow.