@erikdubbelboer 2024-09-22
This is currently not documented at all. I'm thinking I would rather have the function return the correct values even for big numbers than have it be 20ns faster. Just using `strconv.ParseFloat` would also result in less code to maintain. What do you think?