@ksw2000 2024-09-22
Parse the float string is a complex topic. In `strconv.ParseFloat`, they use IEEE754 unbiased rounding.
In our implementation, we directly truncate the digits after the 19th position.
Although our implementation doesn't achieve the same high precision as `strconv.ParseFloat`, the error is within an acceptable range for network applications.
I think we can add the comments that our `ParseUfloat` is much faster than `strconv.ParseFloat`, but the results are less accurate.


```
goos: windows
goarch: amd64
pkg: github.com/valyala/fasthttp
cpu: Intel(R) Core(TM) i5-8250U CPU @ 1.60GHz
BenchmarkParseUfloat-8                  13448752                78.83 ns/op
BenchmarkParseUfloatStrconv-8            4897358               248.1 ns/op
PASS
ok      github.com/valyala/fasthttp     2.738s
```

```
1234e23
ParseUfloat:        1.234e+26
strconv.ParseFloat: 1.2339999999999998e+26

1.234e+32
ParseUfloat:        1.234e+32
strconv.ParseFloat: 1.2339999999999998e+32

123456789123456789.987654321
ParseUfloat:        1.2345678912345678e+17
strconv.ParseFloat: 1.2345678912345678e+17
```