@erikdubbelboer 2024-09-22
There are other inconsistencies as well. For example this test fails because it returns `0` instead of `1` like `strconv.ParseFloat` does.
```
testParseUfloatSuccess(t, "00000000000000000001", 1)
```

Or parsing `87000000000.2` results in `87000000000.200012` while `strconv.ParseFloat` returns `87000000000.199997` which is a bigger difference for such a low number than I would like.