@ksw2000 2024-09-22
In the first case `00 ... 001`, we can add some code to remove the leading zeros. As for the second case `87000000000.2`, the relative error between `ParseUfloat` and `strconv.ParseFloat` is only **1.75e-14**.