@ksw2000 2024-09-22
The second case is very strange. Our algorithm calculates using `float64(870000000002) * math.Pow10(-1)`, but if we change to using division, the result will be the same as `strconv.ParseFloat`.

```go
fmt.Printf("%f %f\n", float64(870000000002)*math.Pow10(-1), float64(870000000002)/math.Pow10(1))
```

```
87000000000.200012 87000000000.199997
```