@erikdubbelboer 2024-09-22
`71000000000000.1` results in a `1.562500e-02` relative difference. I'm trying to fuzz test the implementations against each other and I keep finding differences. There is also how they handle numbers with `E` or `X`. Or `strconv.ParseFloat` sometimes returning `NaN` instead of an error. There are so many differences that I really don't think it's worth the 20ns (nanoseconds) overhead. Unless someone is calling this function billions of times it's not going to make a difference. And if someone does that it's much better for them to use a more specific version in their own codebase.