@ksw2000 2024-09-22
The example you provided, using `Relative Error = (expected value - actual value) / actual value`, the error is indeed very small, all below **10^-15**. However, if our function cannot perfectly handle special cases like NaN, I agree with your suggestion. Perhaps we should consider directly calling `strconv.ParseFloat`, or rewriting our function based on `strconv.ParseFloat`. What do you think?