@erikdubbelboer 2024-09-22
Lets just call `strconv.ParseFloat` and include some error handling for negative numbers to keep that backwards compatible.